apps/auth-service/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Yura Mishin",email = "robothot@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.118.3,<0.119.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "dynaconf (>=3.2.12,<4.0.0)",
    "pydantic[email] (>=2.12.0,<3.0.0)",
    "uvicorn (>=0.37.0,<0.38.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
